+++ src/Shared/DataType/ComponentFiltersInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace OxidEsales\GraphQL\ConfigurationAccess\Shared\DataType;
-
-interface ComponentFiltersInterface
-{
-    public function filterComponent(ComponentDataTypeInterface $component): bool;
-}